apps/opportunity_finder/ingestion_service/requirements.txt
scrapy==2.11.0
playwright==1.40.0
kafka-python==2.0.2
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
asyncio-throttle==1.0.2
feedparser==6.0.10
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
aiohttp==3.9.1